Refactor/filter: 폴더정리(Filters폴더 생성하여 폴더 컴포넌트를 모아둠)
