+++ src/components/Filter/MultiRangeSlider/MultiRangeSlider.js
