+++ src/components/Filter/MultiRangeSlider/MultiRangeSlider.module.scss
